Fixed castling * Now able to castle * Fixed Square type requiring cast issue => Still need to make castling more rubost and cancel the move

+++ tests/move/tests_gen_move.cpp
@@ -15,50 +15,50 @@ protected:
 
 TEST_F(MoveGenTest, MoveGenKing)
 {
-    ASSERT_EQ(MoveGen::KingAttacks[static_cast<int>(Square::d4)], 0b1110000010100000111000000000000000000ULL)
+    ASSERT_EQ(MoveGen::KingAttacks[Square::d4], 0b1110000010100000111000000000000000000ULL)
         << "Erreur : Les mouvements du roi sont invalides";
 
-    const U64 bitboard_moves = MoveGen::KingAttacks[static_cast<int>(Square::b3)];
+    const U64 bitboard_moves = MoveGen::KingAttacks[Square::b3];
     ASSERT_EQ(bitboard_moves, 0x7050700);
 }
 
 TEST_F(MoveGenTest, MaskGenRook)
 {
-    ASSERT_EQ(MoveGen::RookMasks[static_cast<int>(Square::d4)], 0b1000000010000000100001110110000010000000100000000000ULL)
+    ASSERT_EQ(MoveGen::RookMasks[Square::d4], 0b1000000010000000100001110110000010000000100000000000ULL)
         << "Erreur : Le mask de lookup de la tour est invalide";
 }
 
 TEST_F(MoveGenTest, MaskGenBishop)
 {
-    ASSERT_EQ(MoveGen::BishopMasks[static_cast<int>(Square::d4)], 0b1000000001000100001010000000000000101000010001000000000ULL)
+    ASSERT_EQ(MoveGen::BishopMasks[Square::d4], 0b1000000001000100001010000000000000101000010001000000000ULL)
         << "Erreur : Le mask de lookup du fou est invalide";
 }
 
 TEST_F(MoveGenTest, MoveGenPawn)
 {
-    ASSERT_EQ(MoveGen::PawnAttacksWhite[static_cast<int>(Square::d4)], 0b1010000000000000000000000000000000000ULL)
+    ASSERT_EQ(MoveGen::PawnAttacksWhite[Square::d4], 0b1010000000000000000000000000000000000ULL)
         << "Erreur : Le mask d'attaque du pion blanc est invalide";
-    ASSERT_EQ(MoveGen::PawnAttacksBlack[static_cast<int>(Square::d4)], 0b101000000000000000000ULL)
+    ASSERT_EQ(MoveGen::PawnAttacksBlack[Square::d4], 0b101000000000000000000ULL)
         << "Erreur : Le mask d'attaque du pion noir est invalide";
 
-    ASSERT_EQ(MoveGen::PawnPushWhite[static_cast<int>(Square::d4)], 0b100000000000000000000000000000000000ULL)
+    ASSERT_EQ(MoveGen::PawnPushWhite[Square::d4], 0b100000000000000000000000000000000000ULL)
         << "Erreur : Le mask de push du pion blanc est invalide";
-    ASSERT_EQ(MoveGen::PawnPushBlack[static_cast<int>(Square::d4)], 0b10000000000000000000ULL)
+    ASSERT_EQ(MoveGen::PawnPushBlack[Square::d4], 0b10000000000000000000ULL)
         << "Erreur : Le mask de push du pion noir est invalide";
 
-    ASSERT_EQ(MoveGen::PawnPush2White[static_cast<int>(Square::d4)], 0b0ULL)
+    ASSERT_EQ(MoveGen::PawnPush2White[Square::d4], 0b0ULL)
         << "Erreur : Le mask de double push du pion blanc est invalide";
-    ASSERT_EQ(MoveGen::PawnPush2Black[static_cast<int>(Square::d4)], 0b0ULL)
+    ASSERT_EQ(MoveGen::PawnPush2Black[Square::d4], 0b0ULL)
         << "Erreur : Le mask de double push du pion noir est invalide";
 
-    ASSERT_EQ(MoveGen::PawnPush2White[static_cast<int>(Square::d2)], 0b1000000000000000000000000000ULL)
+    ASSERT_EQ(MoveGen::PawnPush2White[Square::d2], 0b1000000000000000000000000000ULL)
         << "Erreur : Le mask de double push du pion blanc est invalide";
-    ASSERT_EQ(MoveGen::PawnPush2Black[static_cast<int>(Square::d2)], 0b0ULL)
+    ASSERT_EQ(MoveGen::PawnPush2Black[Square::d2], 0b0ULL)
         << "Erreur : Le mask de double push du pion noir est invalide";
 
-    ASSERT_EQ(MoveGen::PawnPush2White[static_cast<int>(Square::d7)], 0b0ULL)
+    ASSERT_EQ(MoveGen::PawnPush2White[Square::d7], 0b0ULL)
         << "Erreur : Le mask de double push du pion blanc est invalide";
-    ASSERT_EQ(MoveGen::PawnPush2Black[static_cast<int>(Square::d7)], 0b100000000000000000000000000000000000ULL)
+    ASSERT_EQ(MoveGen::PawnPush2Black[Square::d7], 0b100000000000000000000000000000000000ULL)
         << "Erreur : Le mask de double push du pion noir est invalide";
 }
 
@@ -68,7 +68,7 @@ TEST_F(MoveGenTest, MoveGenRook)
     b.load_fen("8/1p1q4/5k2/1n1R4/8/1K2N1p1/3B4/8 w - - 0 1");
 
     // Half legal move (can capture his own pieces)
-    const U64 bitboard_moves = MoveGen::generate_rook_moves(static_cast<int>(Square::d5), b);
+    const U64 bitboard_moves = MoveGen::generate_rook_moves(Square::d5, b);
     ASSERT_EQ(bitboard_moves, 0b1000000010001111011000001000000010000000100000000000);
 }
 
@@ -78,7 +78,7 @@ TEST_F(MoveGenTest, MoveGenRookPieceOnBorder)
     b.load_fen("8/1p1q4/5k2/1n1R3r/8/1K2N1p1/3B4/8 w - - 0 1");
 
     // Half legal move (can capture his own pieces)
-    const U64 bitboard_moves = MoveGen::generate_rook_moves(static_cast<int>(Square::d5), b);
+    const U64 bitboard_moves = MoveGen::generate_rook_moves(Square::d5, b);
     ASSERT_EQ(bitboard_moves, 0b1000000010001111011000001000000010000000100000000000);
 }
 
@@ -88,6 +88,6 @@ TEST_F(MoveGenTest, MoveGenBishop)
     b.load_fen("8/1p1q4/5k2/1n1R4/8/1K2N1p1/3B4/8 w - - 0 1");
 
     // Half legal move (can capture his own pieces)
-    const U64 bitboard_moves = MoveGen::generate_bishop_moves(static_cast<int>(Square::d2), b);
+    const U64 bitboard_moves = MoveGen::generate_bishop_moves(Square::d2, b);
     ASSERT_EQ(bitboard_moves, 0b100000010000101000000000000010100);
 }
